Create updatedmain.c
set_row can light up a row given the RGB in the bottom three bits of a uint8_t. the second argument is the row# you want to light up and the last is the current row, row. 
set_col is the same but with columns. unfortunately, funky things happen when you light up multiple columns or multiple rows (in the same upper or lower part of the matrix)